Selectable text minigame complete(?)
Scoring updates in the top right corner of the screen, ending the game prompts a debug message, and the top left corner of the popup appears at your mouse position, unless it would go off the screen, in which case it gets clamped to the side of the screen.
also, getting to the answer form will now automatically select the inputfield, and pressing enter will allow you to submit your answer, as well as the button.

TODO: highlight edited text, add graphics