Minigame1 Menu changes
Minigame1 has had its menu adjusted so that minimenus show up in appropriate places on the screen, and the main image has the option to either stretch to fit the screen, or scale to fit the screen vertically while leaving a black box outside.

As an added note, the menu for minigame 2 also now has the option to add the black border.