Mushroom man and Dialogue in Overworld
Mushroom man has been added to the map! So far he does not rotate, but I think I will add that once keyboard controls are implemented (there may be some more changes to make before rotation).

Also, a camera cutscene has been added when clicking on characters in the map! There is a blank spot in the code where the dialogue event will be.
Furthermore, clicking on objects in the map now makes it impossible to click other objects until the camera animation is done.