Minigame2 Additions
Added a file reader for content that randomly picks a story for Minigame2. Also added the content files, which have mistakes built into them with a revised system.

Furthermore, a font menu now exists to the left of the scroll. Bold can be toggled with the B button, and the fancy font (which has now also been fixed visually) can. be toggled on and off using the F button. The arrows to the right and left of the number can be used to adjust the font size, or a number can be typed in. Font sizes range from 20 - 75 in this version, and start at 24 by default.