Minigame3 Info Menu and win screen
Copied info screen into minigame3, and messed with priority of menus. Also added a Win screen, and got the game to reset when the quit button is pressed. A minor bug to note is that the pause menu appears over the win screen in minigame1. This can be fixed, but it might not be worth the time.